Former-commit-id: d18d6935718da1ec40b474fe932d9934022f490c
Former-commit-id: 9fdf7a988226c36a262eb7c70994509153517910

+++ server/src/main/java/org/icgc/dcc/web/NextReleaseResource.java
@@ -5,15 +5,19 @@ import java.util.List;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
+import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Request;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
+import javax.ws.rs.core.SecurityContext;
 
 import org.icgc.dcc.release.NextRelease;
 import org.icgc.dcc.release.ReleaseService;
 import org.icgc.dcc.release.model.Release;
+import org.icgc.dcc.shiro.AuthorizationPrivileges;
+import org.icgc.dcc.shiro.ShiroSecurityContext;
 
 import com.google.inject.Inject;
 
@@ -30,11 +34,15 @@ public class NextReleaseResource {
   }
 
   @POST
-  public Response release(Release nextRelease, @Context Request req) {
+  public Response release(Release nextRelease, @Context Request req, @Context SecurityContext securityContext) {
+    if(((ShiroSecurityContext) securityContext).getSubject().isPermitted(
+        AuthorizationPrivileges.RELEASE_CLOSE.toString()) == false) {
+      return Response.status(Status.UNAUTHORIZED).entity(new ServerErrorResponseMessage("Unauthorized")).build();
+    }
     NextRelease oldRelease = releaseService.getNextRelease();
     // Check the timestamp of the oldRelease, since that is the object being updated
     ResponseTimestamper.evaluate(req, oldRelease.getRelease());
-    NextRelease newRelease = oldRelease.release(nextRelease);
+    NextRelease newRelease = oldRelease.release(nextRelease.getName());
 
     return ResponseTimestamper.ok(newRelease.getRelease()).build();
   }
@@ -49,20 +57,30 @@ public class NextReleaseResource {
 
   @POST
   @Path("queue")
-  public Response queue(List<String> projectKeys, @Context Request req) {
+  public Response queue(List<String> projectKeys, @Context Request req, @Context SecurityContext securityContext) {
+    for(String projectKey : projectKeys) {
+      if(((ShiroSecurityContext) securityContext).getSubject().isPermitted(
+          AuthorizationPrivileges.projectViewPrivilege(projectKey)) == false) {
+        return Response.status(Status.UNAUTHORIZED).entity(new ServerErrorResponseMessage("Unauthorized")).build();
+      }
+    }
     ResponseTimestamper.evaluate(req, this.releaseService.getNextRelease().getRelease());
 
     if(this.releaseService.hasProjectKey(projectKeys)) {
       this.releaseService.queue(projectKeys);
       return Response.ok().build();
     } else {
-      return Response.status(Status.BAD_REQUEST).build();
+      return Response.status(Status.BAD_REQUEST).entity(new ServerErrorResponseMessage("ProjectKeyNotFound")).build();
     }
   }
 
   @DELETE
   @Path("queue")
-  public Response removeAllQueued() {
+  public Response removeAllQueued(@Context SecurityContext securityContext) {
+    if(((ShiroSecurityContext) securityContext).getSubject().isPermitted(
+        AuthorizationPrivileges.QUEUE_DELETE.toString()) == false) {
+      return Response.status(Status.UNAUTHORIZED).entity(new ServerErrorResponseMessage("Unauthorized")).build();
+    }
     this.releaseService.deleteQueuedRequest();
 
     return Response.ok().build();
@@ -78,12 +96,37 @@ public class NextReleaseResource {
 
   @POST
   @Path("signed")
-  public Response signOff(List<String> projectKeys, @Context Request req) {
+  public Response signOff(List<String> projectKeys, @Context Request req, @Context SecurityContext securityContext) {
+    if(((ShiroSecurityContext) securityContext).getSubject().isPermitted(
+        AuthorizationPrivileges.RELEASE_SIGNOFF.toString()) == false) {
+      return Response.status(Status.UNAUTHORIZED).entity(new ServerErrorResponseMessage("Unauthorized")).build();
+    }
     ResponseTimestamper.evaluate(req, this.releaseService.getNextRelease().getRelease());
 
     if(this.releaseService.hasProjectKey(projectKeys)) {
       this.releaseService.signOff(projectKeys);
       return Response.ok().build();
+    } else {
+      return Response.status(Status.BAD_REQUEST).entity(new ServerErrorResponseMessage("ProjectKeyNotFound")).build();
+    }
+  }
+
+  @PUT
+  @Path("update")
+  public Response update(Release release, @Context Request req, @Context SecurityContext securityContext) {
+    if(((ShiroSecurityContext) securityContext).getSubject().isPermitted(
+        AuthorizationPrivileges.RELEASE_MODIFY.toString()) == false) {
+      return Response.status(Status.UNAUTHORIZED).entity(new ServerErrorResponseMessage("Unauthorized")).build();
+    }
+    if(release != null) {
+      ResponseTimestamper.evaluate(req, release);
+
+      if(this.releaseService.list().isEmpty()) {
+        return Response.status(Status.BAD_REQUEST).build();
+      } else {
+        Release updatedRelease = releaseService.update(release);
+        return ResponseTimestamper.ok(updatedRelease).build();
+      }
     } else {
       return Response.status(Status.BAD_REQUEST).build();
     }
